refactor interface file :car:
