+++ Phase06/p6/db.json
@@ -69,6 +69,13 @@
       "lastName": "sdfsdfsdf",
       "email": "h.nazarieh@gmail.com",
       "password": "sdfsdfsd1f6511"
+    },
+    {
+      "id": "96ae",
+      "firstName": "z",
+      "lastName": "z",
+      "email": "z@2",
+      "password": "z@2222222"
     }
   ],
   "books": [
@@ -76,7 +83,10 @@
       "id": "4871",
       "name": "test",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Fiction", " Philosophy"],
+      "genre": [
+        "Fiction",
+        " Philosophy"
+      ],
       "author": "Paulo Coelho",
       "publishData": "1988-01-01",
       "price": 1099
@@ -84,7 +94,10 @@
     {
       "name": "Brave New World",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Science Fiction", "Dystopian"],
+      "genre": [
+        "Science Fiction",
+        "Dystopian"
+      ],
       "author": "Aldous Huxley",
       "publishData": "1932-01-01",
       "price": 899,
@@ -93,7 +106,10 @@
     {
       "name": "The Hunger Games",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Young Adult", "Science Fiction"],
+      "genre": [
+        "Young Adult",
+        "Science Fiction"
+      ],
       "author": "Suzanne Collins",
       "publishData": "2008-09-14",
       "price": 1199,
@@ -102,7 +118,10 @@
     {
       "name": "The Shining",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Horror", "Thriller"],
+      "genre": [
+        "Horror",
+        "Thriller"
+      ],
       "author": "Stephen King",
       "publishData": "1977-01-28",
       "price": 1299,
@@ -111,7 +130,10 @@
     {
       "name": "The Girl with the Dragon Tattoo",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Crime", "Mystery"],
+      "genre": [
+        "Crime",
+        "Mystery"
+      ],
       "author": "Stieg Larsson",
       "publishData": "2005-08-01",
       "price": 1399,
@@ -120,7 +142,10 @@
     {
       "name": "The Hitchhiker's Guide to the Galaxy",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Science Fiction", "Comedy"],
+      "genre": [
+        "Science Fiction",
+        "Comedy"
+      ],
       "author": "Douglas Adams",
       "publishData": "1979-10-12",
       "price": 999,
@@ -129,7 +154,10 @@
     {
       "name": "The Kite Runner",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Fiction", "Historical"],
+      "genre": [
+        "Fiction",
+        "Historical"
+      ],
       "author": "Khaled Hosseini",
       "publishData": "2003-05-29",
       "price": 1199,
@@ -138,7 +166,10 @@
     {
       "name": "The Martian",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Science Fiction", "Adventure"],
+      "genre": [
+        "Science Fiction",
+        "Adventure"
+      ],
       "author": "Andy Weir",
       "publishData": "2011-09-27",
       "price": 1299,
@@ -147,7 +178,10 @@
     {
       "name": "The Handmaid's Tale",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Dystopian", "Speculative Fiction"],
+      "genre": [
+        "Dystopian",
+        "Speculative Fiction"
+      ],
       "author": "Margaret Atwood",
       "publishData": "1985-06-14",
       "price": 1099,
@@ -156,7 +190,10 @@
     {
       "name": "The Fault in Our Stars",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Young Adult", "Romance"],
+      "genre": [
+        "Young Adult",
+        "Romance"
+      ],
       "author": "John Green",
       "publishData": "2012-01-10",
       "price": 999,
@@ -165,7 +202,10 @@
     {
       "name": "Sapiens: A Brief History of Humankind",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Non-fiction", "History"],
+      "genre": [
+        "Non-fiction",
+        "History"
+      ],
       "author": "Yuval Noah Harari",
       "publishData": "2011-01-01",
       "price": 1599,
@@ -174,7 +214,10 @@
     {
       "name": "The Lord of the Rings",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Fantasy", "Epic"],
+      "genre": [
+        "Fantasy",
+        "Epic"
+      ],
       "author": "J.R.R. Tolkien",
       "publishData": "1954-07-29",
       "price": 1799,
@@ -183,7 +226,10 @@
     {
       "name": "Becoming",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Non-fiction", "Autobiography"],
+      "genre": [
+        "Non-fiction",
+        "Autobiography"
+      ],
       "author": "Michelle Obama",
       "publishData": "2018-11-13",
       "price": 1499,
@@ -193,7 +239,10 @@
       "id": "6368",
       "name": "One Hundred Years of Solitude",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Magical Realism", " Literary Fiction"],
+      "genre": [
+        "Magical Realism",
+        " Literary Fiction"
+      ],
       "author": "Gabriel García Márquez",
       "publishData": "1967-05-30",
       "price": 1099
@@ -201,7 +250,10 @@
     {
       "name": "The Silent Patient",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Thriller", "Mystery"],
+      "genre": [
+        "Thriller",
+        "Mystery"
+      ],
       "author": "Alex Michaelides",
       "publishData": "2019-02-05",
       "price": 1399,
@@ -210,7 +262,10 @@
     {
       "name": "The Catcher in the Rye",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Fiction", "Coming-of-age"],
+      "genre": [
+        "Fiction",
+        "Coming-of-age"
+      ],
       "author": "J.D. Salinger",
       "publishData": "1951-07-16",
       "price": 999,
@@ -219,7 +274,10 @@
     {
       "name": "Educated",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Non-fiction", "Memoir"],
+      "genre": [
+        "Non-fiction",
+        "Memoir"
+      ],
       "author": "Tara Westover",
       "publishData": "2018-02-20",
       "price": 1299,
@@ -228,7 +286,10 @@
     {
       "name": "The Pillars of the Earth",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Historical Fiction", "Epic"],
+      "genre": [
+        "Historical Fiction",
+        "Epic"
+      ],
       "author": "Ken Follett",
       "publishData": "1989-10-01",
       "price": 1599,
@@ -237,7 +298,10 @@
     {
       "name": "The Immortal Life of Henrietta Lacks",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Non-fiction", "Science"],
+      "genre": [
+        "Non-fiction",
+        "Science"
+      ],
       "author": "Rebecca Skloot",
       "publishData": "2010-02-02",
       "price": 1199,
@@ -246,7 +310,10 @@
     {
       "name": "The name of the Wind",
       "image": "https://picsum.photos/200/300",
-      "genre": ["Fantasy", "Adventure"],
+      "genre": [
+        "Fantasy",
+        "Adventure"
+      ],
       "author": "Patrick Rothfuss",
       "publishData": "2007-03-27",
       "price": 1499,
@@ -256,7 +323,9 @@
       "id": "ca5e",
       "name": "atal",
       "image": "eqweqwe",
-      "genre": ["qweqw"],
+      "genre": [
+        "qweqw"
+      ],
       "author": "qweqeqwe",
       "publishData": "2024-08-01",
       "price": 32131
