+++ Phase06/p6/src/app/app.component.html
@@ -1,3 +1,4 @@
 <app-header></app-header>
 <router-outlet></router-outlet>
+
 <app-footer></app-footer>
