+++ Phase06/p6/src/app/component/aside/aside.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 import { FormGroup, FormControl } from '@angular/forms';
 import { ReactiveFormsModule, Validators } from '@angular/forms';
 import { ApiService } from '../../service/api.service';
-import Book from '../../interface/book';
+import Book from '../../interface/interface';
 
 @Component({
   selector: 'app-aside',
