+++ Phase06/p6/src/app/component/header/header.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { RouterLink, RouterLinkActive } from '@angular/router';
-import { User } from '../../interface/book';
+import { User } from '../../interface/interface';
 import { UserService } from '../../service/user.service';
 
 @Component({
