+++ Phase06/p6/src/app/component/home/home.component.scss
@@ -13,6 +13,7 @@
     inline-size: 80%;
 
     flex-grow: 1;
+    
     display: flex;
     flex-wrap: wrap;
     flex-direction: row;
