+++ Phase06/p6/src/app/component/main/main.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ApiService } from '../../service/api.service';
 import { BookComponent } from './book/book.component';
-import Book from '../../interface/book';
+import Book from '../../interface/interface';
 
 @Component({
   selector: 'app-main',
