+++ Phase06/p6/src/app/component/signup/signup.component.html
@@ -46,8 +46,8 @@
     />
 
     <small class="text-danger" [hidden]="formGroup.get('password')?.valid">
-      password must be at least 8 characters long. and contain at least one
-      number and one letter.
+      password must be at least 8 characters long.
+      and contain at least one number and one letter.
     </small>
 
     <button type="submit">Sign Up</button>
