+++ Phase06/p6/src/app/guards/authenticator-guard.guard.ts
@@ -1,7 +1,7 @@
 import { CanActivateFn } from '@angular/router';
 import { inject } from '@angular/core';
 import { UserService } from '../service/user.service';
-import { User } from '../interface/book';
+import { User } from '../interface/interface';
 
 export const authenticatorGuardGuard: CanActivateFn = (route, state) => {
   const service = inject(UserService);
