+++ Phase06/p6/src/app/interface/interface.ts
